Updated OLED.txt
Three rules: 
1.	Place a comma before a conjunction introducing an independent clause
2.	The number of the subject determines the number of verb
3.	A participial phrase at the beginning of a sentence must refer to the grammatical subject

According to the marks I made on text, the three rules are applied correctly. The conjunction introducing an independent clause, and there is a comma before it. Since it is wrong to join two independent clauses with a comma, to introduce another independent clause, I added a conjunction after a comma to finish this sentence. The subject "size" is not directly connected to the verb, but by tracing back to the real subject, we are able to determine the number of verbs and transform them into single form.  For the third rule, there is one sentence that meet this criterion. The word “having” refers to OLED, which is also the subject of this sentence, so this participial phrase at the beginning is used correctly.